refactor(database): restructure SQLite integration and move enums to a dedicated file
