+++ src/node/sqlite/enums.sqlite.ts
